+++ argilla-server/src/argilla_server/api/handlers/v1/records.py
@@ -25,7 +25,7 @@ from argilla_server.api.schemas.v1.records import RecordUpdate
 from argilla_server.api.schemas.v1.responses import Response, ResponseCreate
 from argilla_server.api.schemas.v1.suggestions import Suggestion as SuggestionSchema
 from argilla_server.api.schemas.v1.suggestions import SuggestionCreate, Suggestions
-from argilla_server.contexts import datasets
+from argilla_server.contexts import datasets, records
 from argilla_server.database import get_async_db
 from argilla_server.errors.future.base_errors import NotFoundError, UnprocessableEntityError
 from argilla_server.models import Dataset, Question, Record, Suggestion, User
@@ -74,16 +74,21 @@ async def update_record(
         db,
         record_id,
         options=[
-            selectinload(Record.dataset).selectinload(Dataset.questions),
-            selectinload(Record.dataset).selectinload(Dataset.metadata_properties),
+            selectinload(Record.dataset).options(
+                selectinload(Dataset.questions),
+                selectinload(Dataset.metadata_properties),
+                selectinload(Dataset.vectors_settings),
+                selectinload(Dataset.fields),
+            ),
             selectinload(Record.suggestions),
+            selectinload(Record.responses),
             selectinload(Record.vectors),
         ],
     )
 
     await authorize(current_user, RecordPolicy.update(record))
 
-    return await datasets.update_record(db, search_engine, record, record_update)
+    return await records.update_record(db, search_engine, record, record_update)
 
 
 @router.post("/records/{record_id}/responses", status_code=status.HTTP_201_CREATED, response_model=Response)
@@ -233,4 +238,4 @@ async def delete_record(
 
     await authorize(current_user, RecordPolicy.delete(record))
 
-    return await datasets.delete_record(db, search_engine, record)
+    return await records.delete_record(db, search_engine, record)