+++ argilla-server/src/argilla_server/api/schemas/v1/records_bulk.py
@@ -30,7 +30,7 @@ class RecordsBulk(BaseModel):
     items: List[Record]
 
 
-class RecordsBulkWithUpdateInfo(RecordsBulk):
+class RecordsBulkWithUpdatedItemIds(RecordsBulk):
     updated_item_ids: List[UUID]