+++ argilla-server/src/argilla_server/contexts/datasets.py
@@ -12,22 +12,14 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import copy
 from collections import defaultdict
 from datetime import datetime
 from typing import (
     TYPE_CHECKING,
-    Any,
-    Callable,
-    Dict,
     Iterable,
     List,
-    Literal,
     Optional,
     Sequence,
-    Set,
-    Tuple,
-    TypeVar,
     Union,
 )
 from uuid import UUID
@@ -41,9 +33,7 @@ from sqlalchemy.orm import contains_eager, joinedload, selectinload
 from argilla_server.api.schemas.v1.fields import FieldCreate
 from argilla_server.api.schemas.v1.metadata_properties import MetadataPropertyCreate, MetadataPropertyUpdate
 from argilla_server.api.schemas.v1.records import (
-    RecordCreate,
     RecordIncludeParam,
-    RecordUpdateWithId,
 )
 from argilla_server.api.schemas.v1.responses import (
     ResponseCreate,
@@ -101,7 +91,6 @@ from argilla_server.validators.suggestions import SuggestionCreateValidator
 
 if TYPE_CHECKING:
     from argilla_server.api.schemas.v1.fields import FieldUpdate
-    from argilla_server.api.schemas.v1.records import RecordUpdate
     from argilla_server.api.schemas.v1.suggestions import SuggestionCreate
     from argilla_server.api.schemas.v1.vector_settings import VectorSettingsUpdate
 
@@ -333,6 +322,7 @@ async def create_vector_settings(
     return vector_settings
 
 
+# TODO: Move this function to the records.py context
 async def get_records_by_ids(
     db: AsyncSession,
     records_ids: Iterable[UUID],
@@ -462,103 +452,6 @@ async def get_dataset_users_progress(db: AsyncSession, dataset: Dataset) -> List
     return [{"username": username, **progress} for username, progress in annotators_progress.items()]
 
 
-_EXTRA_METADATA_FLAG = "extra"
-
-
-async def _validate_metadata(
-    db: AsyncSession,
-    dataset: Dataset,
-    metadata: Dict[str, Any],
-    metadata_properties: Optional[Dict[str, Union[MetadataProperty, Literal["extra"]]]] = None,
-) -> Dict[str, Union[MetadataProperty, Literal["extra"]]]:
-    if metadata_properties is None:
-        metadata_properties = {}
-
-    for name, value in metadata.items():
-        metadata_property = metadata_properties.get(name)
-
-        if metadata_property is None:
-            metadata_property = await MetadataProperty.get_by(db, name=name, dataset_id=dataset.id)
-
-            # If metadata property does not exists but extra metadata is allowed, then we set a flag value to
-            # avoid querying the database again
-            if metadata_property is None and dataset.allow_extra_metadata:
-                metadata_property = _EXTRA_METADATA_FLAG
-                metadata_properties[name] = metadata_property
-            elif metadata_property is not None:
-                metadata_properties[name] = metadata_property
-            else:
-                raise ValueError(
-                    f"'{name}' metadata property does not exists for dataset '{dataset.id}' and extra metadata is"
-                    " not allowed for this dataset"
-                )
-
-        # If metadata property is not found and extra metadata is allowed, then we skip the value validation
-        if metadata_property == _EXTRA_METADATA_FLAG:
-            continue
-
-        try:
-            if value is not None:
-                metadata_property.parsed_settings.check_metadata(value)
-        except (UnprocessableEntityError, ValueError) as e:
-            raise UnprocessableEntityError(f"'{name}' metadata property validation failed because {e}") from e
-
-    return metadata_properties
-
-
-async def validate_user_exists(db: AsyncSession, user_id: UUID, users_ids: Optional[Set[UUID]]) -> Set[UUID]:
-    if not users_ids:
-        users_ids = set()
-
-    if user_id not in users_ids:
-        if not await accounts.user_exists(db, user_id):
-            raise UnprocessableEntityError(f"user_id={str(user_id)} does not exist")
-
-        users_ids.add(user_id)
-
-    return users_ids
-
-
-async def _validate_vector(
-    db: AsyncSession,
-    dataset_id: UUID,
-    vector_name: str,
-    vector_value: List[float],
-    vectors_settings: Optional[Dict[str, VectorSettingsSchema]] = None,
-) -> Dict[str, VectorSettingsSchema]:
-    if vectors_settings is None:
-        vectors_settings = {}
-
-    vector_settings = vectors_settings.get(vector_name, None)
-    if not vector_settings:
-        vector_settings = await VectorSettings.get_by(db, name=vector_name, dataset_id=dataset_id)
-        if not vector_settings:
-            raise UnprocessableEntityError(
-                f"vector with name={str(vector_name)} does not exist for dataset_id={str(dataset_id)}"
-            )
-
-        vector_settings = VectorSettingsSchema.model_validate(vector_settings)
-        vectors_settings[vector_name] = vector_settings
-
-    vector_settings.check_vector(vector_value)
-
-    return vectors_settings
-
-
-async def _build_record(
-    db: AsyncSession, dataset: Dataset, record_create: RecordCreate, caches: Dict[str, Any]
-) -> Record:
-    _validate_record_fields(dataset, fields=record_create.fields)
-    await _validate_record_metadata(db, dataset, record_create.metadata, caches["metadata_properties_cache"])
-
-    return Record(
-        fields=jsonable_encoder(record_create.fields),
-        metadata_=record_create.metadata,
-        external_id=record_create.external_id,
-        dataset=dataset,
-    )
-
-
 async def _load_users_from_responses(responses: Union[Response, Iterable[Response]]) -> None:
     if isinstance(responses, Response):
         responses = [responses]
@@ -569,162 +462,6 @@ async def _load_users_from_responses(responses: Union[Response, Iterable[Respons
         await response.awaitable_attrs.user
 
 
-async def _validate_record_metadata(
-    db: AsyncSession,
-    dataset: Dataset,
-    metadata: Optional[Dict[str, Any]] = None,
-    cache: Dict[str, Union[MetadataProperty, Literal["extra"]]] = {},
-) -> Dict[str, Union[MetadataProperty, Literal["extra"]]]:
-    """Validate metadata for a record."""
-    if not metadata:
-        return cache
-
-    try:
-        cache = await _validate_metadata(db, dataset=dataset, metadata=metadata, metadata_properties=cache)
-        return cache
-    except (UnprocessableEntityError, ValueError) as e:
-        raise UnprocessableEntityError(f"metadata is not valid: {e}") from e
-
-
-async def _build_record_suggestions(
-    db: AsyncSession,
-    record: Record,
-    suggestions_create: Optional[List["SuggestionCreate"]],
-    questions_cache: Optional[Dict[UUID, Question]] = None,
-) -> List[Suggestion]:
-    """Create suggestions for a record."""
-    if not suggestions_create:
-        return []
-
-    suggestions = []
-    for suggestion_create in suggestions_create:
-        try:
-            if not questions_cache:
-                questions_cache = {}
-
-            question = questions_cache.get(suggestion_create.question_id, None)
-            if not question:
-                question = await Question.get(
-                    db, suggestion_create.question_id, options=[selectinload(Question.dataset)]
-                )
-                if not question:
-                    raise UnprocessableEntityError(f"question_id={str(suggestion_create.question_id)} does not exist")
-                questions_cache[suggestion_create.question_id] = question
-
-            SuggestionCreateValidator.validate(suggestion_create, question.parsed_settings, record)
-
-            suggestions.append(
-                Suggestion(
-                    type=suggestion_create.type,
-                    score=suggestion_create.score,
-                    value=jsonable_encoder(suggestion_create.value),
-                    agent=suggestion_create.agent,
-                    question_id=suggestion_create.question_id,
-                    record=record,
-                )
-            )
-
-        except (UnprocessableEntityError, ValueError) as e:
-            raise UnprocessableEntityError(
-                f"suggestion for question_id={suggestion_create.question_id} is not valid: {e}"
-            ) from e
-
-    return suggestions
-
-
-VectorClass = TypeVar("VectorClass")
-
-
-async def _build_record_vectors(
-    db: AsyncSession,
-    dataset: Dataset,
-    vectors_dict: Dict[str, List[float]],
-    build_vector_func: Callable[[List[float], UUID], VectorClass],
-    cache: Optional[Dict[str, VectorSettingsSchema]] = None,
-) -> List[VectorClass]:
-    """Create vectors for a record."""
-    if not vectors_dict:
-        return []
-
-    vectors = []
-    for vector_name, vector_value in vectors_dict.items():
-        try:
-            cache = await _validate_vector(db, dataset.id, vector_name, vector_value, vectors_settings=cache)
-            vectors.append(build_vector_func(vector_value, cache[vector_name].id))
-        except (UnprocessableEntityError, ValueError) as e:
-            raise UnprocessableEntityError(f"vector with name={vector_name} is not valid: {e}") from e
-
-    return vectors
-
-
-async def _exists_records_with_ids(db: AsyncSession, dataset_id: UUID, records_ids: List[UUID]) -> List[UUID]:
-    result = await db.execute(select(Record.id).filter(Record.dataset_id == dataset_id, Record.id.in_(records_ids)))
-    return result.scalars().all()
-
-
-async def _build_record_update(
-    db: AsyncSession, record: Record, record_update: "RecordUpdateWithId", caches: Optional[Dict[str, Any]] = None
-) -> Tuple[Dict[str, Any], Union[List[Suggestion], None], List[VectorSchema], bool, Dict[str, Any]]:
-    if caches is None:
-        caches = {
-            "metadata_properties": {},
-            "questions": {},
-            "vector_settings": {},
-        }
-
-    params = record_update.model_dump(exclude_unset=True)
-    needs_search_engine_update = False
-    suggestions = None
-    vectors = []
-
-    if "metadata_" in params:
-        metadata = params["metadata_"]
-        needs_search_engine_update = True
-        if metadata is not None:
-            caches["metadata_properties"] = await _validate_record_metadata(
-                db, record.dataset, metadata, caches["metadata_properties"]
-            )
-
-    if record_update.suggestions is not None:
-        params.pop("suggestions")
-        questions_ids = [suggestion.question_id for suggestion in record_update.suggestions]
-        if len(questions_ids) != len(set(questions_ids)):
-            raise UnprocessableEntityError("found duplicate suggestions question IDs")
-        suggestions = await _build_record_suggestions(db, record, record_update.suggestions, caches["questions"])
-
-    if record_update.vectors is not None:
-        params.pop("vectors")
-        vectors = await _build_record_vectors(
-            db,
-            record.dataset,
-            record_update.vectors,
-            build_vector_func=lambda value, vector_settings_id: VectorSchema(
-                value=value, record_id=record_update.id, vector_settings_id=vector_settings_id
-            ),
-            cache=caches["vector_settings"],
-        )
-        needs_search_engine_update = True
-
-    return params, suggestions, vectors, needs_search_engine_update, caches
-
-
-async def _preload_records_relationships_before_index(db: AsyncSession, records: List[Record]) -> None:
-    for record in records:
-        await _preload_record_relationships_before_index(db, record)
-
-
-async def _preload_record_relationships_before_index(db: AsyncSession, record: Record) -> None:
-    await db.execute(
-        select(Record)
-        .filter_by(id=record.id)
-        .options(
-            selectinload(Record.responses).selectinload(Response.user),
-            selectinload(Record.suggestions).selectinload(Suggestion.question),
-            selectinload(Record.vectors),
-        )
-    )
-
-
 async def preload_records_relationships_before_validate(db: AsyncSession, records: List[Record]) -> None:
     await db.execute(
         select(Record)
@@ -733,70 +470,6 @@ async def preload_records_relationships_before_validate(db: AsyncSession, record
             selectinload(Record.dataset).selectinload(Dataset.questions),
         )
     )
-
-
-async def delete_records(
-    db: AsyncSession, search_engine: "SearchEngine", dataset: Dataset, records_ids: List[UUID]
-) -> None:
-    params = [Record.id.in_(records_ids), Record.dataset_id == dataset.id]
-
-    records = (await db.execute(select(Record).filter(*params).order_by(Record.inserted_at.asc()))).scalars().all()
-
-    deleted_record_events_v1 = []
-    for record in records:
-        deleted_record_events_v1.append(
-            await build_record_event_v1(db, RecordEvent.deleted, record),
-        )
-
-    records = await Record.delete_many(db, conditions=params)
-
-    await search_engine.delete_records(dataset=dataset, records=records)
-
-    for deleted_record_event_v1 in deleted_record_events_v1:
-        await deleted_record_event_v1.notify(db)
-
-
-async def update_record(
-    db: AsyncSession, search_engine: "SearchEngine", record: Record, record_update: "RecordUpdate"
-) -> Record:
-    params, suggestions, vectors, needs_search_engine_update, _ = await _build_record_update(
-        db, record, RecordUpdateWithId(id=record.id, **record_update.model_dump(by_alias=True, exclude_unset=True))
-    )
-
-    # Remove existing suggestions
-    if suggestions is not None:
-        record.suggestions = []
-        params["suggestions"] = suggestions
-
-    async with db.begin_nested():
-        record = await record.update(db, **params, replace_dict=True, autocommit=False)
-
-        if vectors:
-            await Vector.upsert_many(
-                db, objects=vectors, constraints=[Vector.record_id, Vector.vector_settings_id], autocommit=False
-            )
-            await db.refresh(record, attribute_names=["vectors"])
-
-    await db.commit()
-
-    if needs_search_engine_update:
-        await _preload_record_relationships_before_index(db, record)
-        await search_engine.index_records(record.dataset, [record])
-
-    await notify_record_event_v1(db, RecordEvent.updated, record)
-
-    return record
-
-
-async def delete_record(db: AsyncSession, search_engine: "SearchEngine", record: Record) -> Record:
-    deleted_record_event_v1 = await build_record_event_v1(db, RecordEvent.deleted, record)
-
-    record = await record.delete(db)
-
-    await search_engine.delete_records(dataset=record.dataset, records=[record])
-    await deleted_record_event_v1.notify(db)
-
-    return record
 
 
 async def create_response(
@@ -917,22 +590,6 @@ async def delete_response(db: AsyncSession, search_engine: SearchEngine, respons
     await deleted_response_event_v1.notify(db)
 
     return response
-
-
-def _validate_record_fields(dataset: Dataset, fields: Dict[str, Any]):
-    fields_copy = copy.copy(fields or {})
-    for field in dataset.fields:
-        if field.required and not (field.name in fields_copy and fields_copy.get(field.name) is not None):
-            raise UnprocessableEntityError(f"missing required value for field: {field.name!r}")
-
-        value = fields_copy.pop(field.name, None)
-        if value and not isinstance(value, str):
-            raise UnprocessableEntityError(
-                f"wrong value found for field {field.name!r}. Expected {str.__name__!r}, found {type(value).__name__!r}"
-            )
-
-    if fields_copy:
-        raise UnprocessableEntityError(f"found fields values for non configured fields: {list(fields_copy.keys())!r}")
 
 
 async def _preload_suggestion_relationships_before_index(db: AsyncSession, suggestion: Suggestion) -> None: